requirements.txt
Django==3.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
pytz==2018.5
django-heroku==0.3.1
django-extensions==3.0.5
